Fix aliasing in main menu audio
Audio play initiation was being run in the run method of the Game instance, which is looped. This caused the main menu music to start playing again and again, each time being played slightly after causing the aliasing effect.

Audio play initiation moved to outside main loop and structure built to handle main menu audio playing and stopping when going to and from the main menu.